Add reworked packet generation infrastructure
Add a far more flexible infrastructure for generating packets.
Unit tests define packet templates, which can be encapsulated
within each other as needed.  Each field in every packet header
is individually settable, and fields that are not explicit set
get a default value.  This should make defining packet templates
significantly easier and more readable.

Packet templates can also be defined by taking an existing
template and applying a new subset of fields, which should
make it easy to create a stream of packets.

The infrastructure for matching packets defined in this way is
forthcoming.