Fix calculation of IP length field
When a payload is encapsulated in an IPv4 header, the ip_len
field needs to be updated to reflect the size of the payload.
Fix this by allowing field setters to advertise that they will
have effects on the entire header stack.  When such a field is
applied it will give us a (potentially stateful) functor that
will be applied to every higher-level header.

Use this to define some packet-length related setters that
propagate the current payload size to every header in the
packet.