Make GlobalMock initialization more resilent
Currently getting a GlobalMock instance initialized required the
test writer to instantiate an object in their test that would
implicitly initialize the GlobalMock.  This worked, but if a
test followed an unexpected path or the test writer forget, the
unit test would just crash.  Instead use the new sysunit initializer
API to register an initializer that will automatically initialize
any GlobalMock linked into the test.

A test will now just fail with an unexpected call on a mock
object rather than segfaulting on a NULL pointer, which is
way easier to debug.