Clean up debug printing
Be explicit about the lifetimes of the temporary layer name
objects.  I'm not conviced that the previous code wasn't reaching
into destructed objects.

Also print the outer header before the inner header.  This causes
the packet to be printed from outside in, matching how a packet
is typically written or displayed.