Ensure that inner Ethernet headers propagate their size
If an Ethernet header is embedded in an another protocol, it needs
to inform the outer headers of its size so that (for example)
the ip length can be calculated correctly.  Definite a
DefaultEncapFieldSetter that does this, and use it in all header
templates to ensure that further defaults can be added in the
future and it will apply to all headers.