Change syntax for specifying header layer
The previous syntax for specifying a header layer was not very
compatible with templates.  When used in a template you had to
explicitly call out the fact that you were using a template, like
this:

pkt.template WithHeaderFields<Layer::L3>(...)

Replace this with the following syntax:

pkt.WithHeader(Layer::L3).Fields(...)

This isn't quite as nice to look at, but it reads well and can be
used consistently in or out of a template.