Use a unique_ptr for mbuf lifetime management
Rather than forcing callers of Generate() to manually deal with
mbuf lifetime management, by default wrap mbufs in a unique_ptr
that will automatically free the chain with m_freem().  Add a
GenerateRawMbuf() function that returns a mbuf* that must be
manually managed.  This is useful when the mbuf is going to be
immediately passed to an interface that will consume it.