Implement TCP segmentation
Implement TCP segmentation when calling Next() on a template.  It
is now possible to set an MTU on a header (it can be any layer --
it doesn't have to be a L3 header).  The MTU will propagate down
the stack of headers and the innermost TCP header will account
for segmentation when choosing the next th_seq value in Next().
Similarly, the payload will segment itself to fit into the MTU.

Note: The MTU will not be respected if it is set so low that
some headers would be cut off by the MTU.